# Optimize LLE function